services/kb-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
pgvector
openai
python-dotenv
